chore(preconfirmation-driver): refactor module layout
