+++ packages/taiko-client-rs/crates/preconfirmation-driver/src/rpc/mod.rs
@@ -24,6 +24,7 @@
 
 /// API trait and implementation.
 pub mod api;
+pub(crate) mod node_api;
 /// RPC server implementation.
 pub mod server;
 /// Request and response types.
@@ -35,3 +36,14 @@ pub use types::{
     NodeStatus, PreconfRpcErrorCode, PublishCommitmentRequest, PublishCommitmentResponse,
     PublishTxListRequest, PublishTxListResponse,
 };
+
+#[cfg(test)]
+mod tests {
+    use super::node_api;
+
+    #[test]
+    fn node_api_module_exists() {
+        let _ = core::mem::size_of::<Option<fn()>>();
+        let _ = &node_api::NODE_RPC_API_MARKER;
+    }
+}
